lego_recognition_system/requirements.txt
ultralytics
gradio
torch
numpy
sam2 @ git+https://github.com/facebookresearch/segment-anything-2.git
pillow
pandas
psutil>=5.9
